emacs/requirements.txt
Django==1.7.1
Jinja2==2.7.3
Markdown==2.5.1
MarkupSafe==0.23
Unidecode==0.04.16
argparse==1.2.1
django-bootstrap3==4.11.0
django-bower==5.0.1
django-d3==3.3.11
django-filter==0.8
django-jquery==1.9.1
django-nvd3==0.7.7
django-registration==1.0
django-registration-redux==1.1
django-twitter-bootstrap==3.3.0
djangorestframework==2.4.4
python-nvd3==0.13.5
python-slugify==0.0.9
six==1.8.0
wsgiref==0.1.2
